Move components to folder.
