+++ packages/cli/examples/commands/ErrorCompCommand.js
@@ -15,10 +15,10 @@ function ThrowError() {
   return React.createElement(
     Box,
     { flexDirection: 'column' },
-    // React.createElement(Static, { items }, ({ key }) => React.createElement(Text, { key }, key)),
-    ...items.map(({ key }) =>
-      React.createElement(Box, { key }, React.createElement(Text, {}, key)),
-    ),
+    React.createElement(Static, { items }, ({ key }) => React.createElement(Text, { key }, key)),
+    // ...items.map(({ key }) =>
+    //   React.createElement(Box, { key }, React.createElement(Text, {}, key)),
+    // ),
     React.createElement(Text, {}, 'Content'),
   );
 }
