+++ packages/cli/src/components/LogWriter.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { useStdout, useStderr } from 'ink';
-import LogBuffer from './LogBuffer';
+import LogBuffer from '../LogBuffer';
 
 export interface LogWriterProps {
   errBuffer: LogBuffer;
